+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 BREAKING CHANGES:
 FEATURES:
+* **resources/f5os_qkview:** New resource added for generating and managing qkview diagnostic files
 IMPROVEMENTS:
 
 ## 1.5.0